// src/models/IUserData.ts
import { UserRoles } from '../utils/userRoles.ts';
import { EditRequest, User } from './User.ts';

export interface IUserData {
  _id: string;
  name: string;
  sname: string;
  tname?: string;
  role: UserRoles;
  email?: string;
  country?: string;
  requizits?: number;
  archive?: boolean;
  access?: boolean;
  allgroups?: boolean;
  parentName?: string;
  actualMail?: string;
  nationality?: string;
  city?: string;
  editRequest?: EditRequest[];
  whatsappNumber?: string;
  techTrainer?: boolean;
  techStudents?: boolean;
  programmers?: boolean;
  adminProblems?: boolean;
  admins?: boolean;
  trainerProblems?: boolean;
  trainingProgram?: boolean;
  araratInt?: boolean;
  report?: boolean;
  urgent?: boolean;
  testForNewTrainers?: boolean;
  importantNews?: boolean;
  chatForChecks?: boolean;
  chatForCheck?: boolean;
  armTrainers?: boolean;
  rusTrainers?: boolean;
  engTrainers?: boolean;
  fraTrainers?: boolean;
  gerTrainers?: boolean;
  psycho?: boolean;
  seance?: boolean;
  online?: boolean;
  offline?: boolean;
  testLessonQuestions?: boolean;
  groupForNewTrainers?: boolean;
  removedFromGroupBy?: User;
  open_level?: number[];
}